add cast and crew into separate creditsSlice for better modularity
