+++ src/features/Movies/movieSlice.js
@@ -10,21 +10,13 @@ const movieSlice = createSlice({
       total_results: 0,
     },
     movieDetails: null,
-    cast: [],
-    crew: [],
     loading: true,
     error: null,
-    castLoading: false,
-    crewLoading: false,
-    castError: null,
-    crewError: null,
   },
   reducers: {
     startFetch: (state) => {
       state.loading = true;
       state.movieDetails = null;
-      state.castError = null;
-      state.crewError = null;
     },
     setMovies: (state, { payload: popularMovies }) => {
       state.movies.page = popularMovies.page;
@@ -43,32 +35,6 @@ const movieSlice = createSlice({
     setMovieDetails: (state, { payload: movieDetails }) => {
       state.movieDetails = movieDetails;
     },
-    setCastStart: (state) => {
-      state.castLoading = true;
-      state.castError = null;
-    },
-    setCast: (state, { payload }) => {
-      state.castLoading = false;
-      const { cast } = payload;
-      state.cast = cast;
-    },
-    setCastError: (state, { payload: error }) => {
-      state.castLoading = false;
-      state.castError = error;
-    },
-    setCrewStart: (state) => {
-      state.crewLoading = true;
-      state.crewError = null;
-    },
-    setCrew: (state, { payload }) => {
-      state.crewLoading = false;
-      const { crew } = payload;
-      state.crew = crew;
-    },
-    setCrewError: (state, { payload: error }) => {
-      state.crewLoading = false;
-      state.crewError = error;
-    },
     fetchPopularMovies: () => {},
     fetchMovieDetails: () => {},
     fetchSearchResults: () => {},
@@ -83,15 +49,6 @@ export const selectMovieDetails = (state) =>
 export const selectTotalPages = (state) => selectMovies(state).total_pages;
 export const selectTotalResults = (state) => selectMovies(state).total_results;
 
-export const selectCast = (state) => selectMoviesState(state).cast;
-export const selectCrew = (state) => selectMoviesState(state).crew;
-export const selectCastLoading = (state) =>
-  selectMoviesState(state).castLoading;
-export const selectCrewLoading = (state) =>
-  selectMoviesState(state).crewLoading;
-export const selectCastError = (state) => selectMoviesState(state).castError;
-export const selectCrewError = (state) => selectMoviesState(state).crewError;
-
 export const {
   setMovies,
   startFetch,
@@ -101,12 +58,6 @@ export const {
   fetchPopularMovies,
   fetchMovieDetails,
   fetchSearchResults,
-  setCastStart,
-  setCrewStart,
-  setCast,
-  setCrew,
-  setCastError,
-  setCrewError,
 } = movieSlice.actions;
 
 export default movieSlice.reducer;
